// web/src/lib/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json }
  | Json[]

export interface Database {
// ...
  public: {
    Tables: {
      jobs: {
        Row: {
          created_at: string | null
          deleted_at: string | null
          id: string
          log: Json | null
          messages: Json | null
          output: string | null
          started_at: string | null
          status: string | null
          status_message: string | null
          task: string | null
          updated_at: string | null
        }
        Insert: {
          created_at?: string | null
          deleted_at?: string | null
          id?: string
          log?: Json | null
          messages?: Json | null
          output?: string | null
          started_at?: string | null
          status?: string | null
          status_message?: string | null
          task?: string | null
          updated_at?: string | null
        }
        Update: {
          created_at?: string | null
          deleted_at?: string | null
          id?: string
          log?: Json | null
          messages?: Json | null
          output?: string | null
          started_at?: string | null
          status?: string | null
          status_message?: string | null
          task?: string | null
          updated_at?: string | null
        }
      }
    }
// ...
  }
// ...
}